requirements.txt
astor==0.8.1
async-timeout==4.0.3
auto_gptq==0.7.1
autoflake==2.3.1
av==12.0.0
bitsandbytes==0.43.0
black==24.3.0
bpy==4.3.0
colorama==0.4.6
conda-pack==0.7.1
crcmod==1.7
cuda-python==12.1.0
dataclasses==0.6
decord==0.6.0
deprecation==2.1.0
easydict==1.13
einops-exts==0.0.4
exceptiongroup==1.2.0
fastavro==1.9.4
fasteners==0.19
flake8==7.0.0
Flask==3.0.3
fqdn==1.5.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
gdown==5.2.0
h5py==3.12.1
hdfs==2.7.3
httplib2==0.22.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
igraph==0.11.8
imath==0.0.2
isoduration==20.11.0
jsonlines==4.0.0
jsonpickle==3.0.4
jsonpointer==2.4
jupyter==1.1.1
librosa==0.10.1
lightning==2.3.3
loguru==0.7.2
mosaicml-streaming==0.7.5
nvidia-cuda-nvcc-cu12==12.1.105
nvidia-pyindex==1.0.9
objsize==0.7.0
open3d==0.18.0
opencv-python==4.9.0.80
OpenEXR==3.3.3
optimum==1.18.1
optree==0.14.1
orjson==3.10.0
panda3d-gltf==1.2.1
pdoc3==0.10.0
peft==0.10.0
pip-system-certs==4.0
point-cloud-utils==0.29.5
polyscope==2.3.0
pycocotools==2.0.7
pydot==1.4.2
pymeshfix==0.17.0
pymongo==4.6.3
pyrender==0.1.45
PySocks==1.7.1
pytest==8.1.1
python-pycg==0.9.2
randomname==0.2.1
roma==1.5.1
rootutils==1.0.7
Rtree==1.3.0
sagemaker==2.242.0
scikit-image==0.23.1
sentence-transformers==2.6.1
simplejson==3.19.2
smplx==0.1.28
spconv-cu121==2.3.8
tensorboard==2.16.2
timm==0.9.16
tomli==2.0.1
torchaudio==2.5.1+cu121
uri-template==1.3.0
usort==1.0.8.post1
wandb==0.20.0
webcolors==1.13
webdataset==0.2.86
Werkzeug==3.0.6
xatlas==0.0.9
xformers==0.0.28.post3
# MoGe @ git+https://github.com/microsoft/MoGe.git@a8c37341bc0325ca99b9d57981cc3bb2bd3e255b
